:recycle: Breakup sub components
